chore: add client.go and move http client related code into it
